Editing the displayed date in the modal window and refactoring the class structure.
