+++ justbuild.it-WebApp/src/main/java/justbuild/it/web/app/service/OfferCreationService.java
@@ -4,7 +4,6 @@ import justbuild.it.web.app.entity.Offer;
 import justbuild.it.web.app.repository.OfferFileRepository;
 import org.springframework.stereotype.Service;
 import java.util.List;
-import java.util.stream.Collectors;
 import static justbuild.it.web.app.entity.constants.AppConstants.OFFERS_FILEPATH;
 
 @Service
@@ -31,24 +30,5 @@ class OfferCreationService implements OfferCreationServiceInterface {
                 .mapToLong(Offer::getOfferId)
                 .reduce(0L, Long::max)
                 + 1L;
-    }
-
-    @Override
-    public List<Offer> getOffersList() {
-        return offerFileRepository.getOffersFromJsonFile(OFFERS_FILEPATH);
-    }
-
-    @Override
-    public List<Offer> getOffersListFilteredBySearchValue(String searchValue) {
-        List<Offer> offers = offerFileRepository.getOffersFromJsonFile(OFFERS_FILEPATH);
-
-        return offers.stream()
-                .filter(value -> value.getOfferContent().toLowerCase().contains(searchValue.toLowerCase())
-                        || value.getOfferId().toString().contains(searchValue)
-                        || value.getCity().toLowerCase().contains(searchValue.toLowerCase())
-                        || value.getUser().getCompany().toLowerCase().contains(searchValue.toLowerCase())
-                        || value.getUser().getLastName().toLowerCase().contains(searchValue.toLowerCase())
-                        || value.getServiceCategory().toString().toLowerCase().contains(searchValue.toLowerCase()))
-                .collect(Collectors.toList());
     }
 }
