+++ justbuild.it-WebApp/src/main/java/justbuild/it/web/app/service/OfferCreationServiceInterface.java
@@ -2,7 +2,6 @@ package justbuild.it.web.app.service;
 
 import justbuild.it.web.app.entity.Offer;
 import org.springframework.stereotype.Service;
-import java.util.List;
 
 @Service
 interface OfferCreationServiceInterface {
@@ -10,8 +9,4 @@ interface OfferCreationServiceInterface {
     void addOffer(Offer offer);
 
     Long getNextOfferId();
-
-    List<Offer> getOffersList();
-
-    List<Offer> getOffersListFilteredBySearchValue(String searchValue);
 }
