+++ justbuild.it-WebApp/src/main/java/justbuild/it/web/app/service/OfferSearchingServiceInterface.java
@@ -1,0 +1,10 @@
+package justbuild.it.web.app.service;
+
+import justbuild.it.web.app.entity.Offer;
+import java.util.List;
+
+public interface OfferSearchingServiceInterface {
+    List<Offer> getOffersList();
+
+    List<Offer> getOffersListFilteredBySearchValue(String searchValue);
+}
